// next-shell/src/types/term_form.types.ts
export interface TermFormData {
    supplier: string;
    supplierName: string;
    contactPerson: string;
    contactPersonName: string;
    contractNo: string;
    mfgPartNo: string;
    suppOrderCode: string;
    active: boolean;
    validFrom: string;
    validTo: string;
    purchaseTerm: string;
    purchaseTermLocation: string;
    purchaseTermLocationName: string;
    purchaseSubLocation: string;
    salesTerm: string;
    salesSubLocation: string;
    incoterm: string;
    prodCost: number;
    pkh: number;
    soc: number;
    currency: string;
    exRate: number;
    shipMode: string;
    dimUnit: string;
    length: number;
    width: number;
    height: number;
    weight: number;
    cWeight: number;
    freightType: string;
    freightRate: number;
    fr: number;
    insPercent: number;
    dutyPercent: number;
    excisePercent: number;
    miscTax: number;
    numInBuy: number;
    numInSale: number;
    markup: number;
    stockFeePercent: number;
    wireTT: number;
    customClear: number;
    scc: number;
    zoneRate: number;
    salesPerson: string;
    salesPersonName: string;
    sourcedBy: string;
    sourcedByName: string;
    remark: string;
    leadTime: string;
    moq: string;
    vendorBPA: boolean;
    purchaseUOM: string;
    salesUOM: string;
    spk: number;
    qoc: number;
    updatedBy: string;
    updatedDate: string;
    stockUOM: string;
}
// ...
export interface TermSupplierOption {
    code: string;
    name: string;
}

export interface TermContactOption {
    code: string;
    name: string;
    active: string;
}

export interface TermLocationOption {
    code: string;
    name: string;
    priority: number;
    zoneName: string;
    zoneRate: number;
}
// ...
export type UpdateTermFormData = <K extends keyof TermFormData>(field: K, value: TermFormData[K]) => void;